Help spacing, code cleanup, fix ENTER on password broken with last commit by leon hölzel...
