+++ electron-vite-project/src/components/LightSettings.css
@@ -245,20 +245,7 @@
   margin: 0;
 }
 
-.LightSettingsTitle {
-  position: absolute;
-  width: 50vw;
-  object-fit: contain;
-  display: flex;
-  font-size: 20px;
-}
-
 .LightSettingsSubTitle {
-  font-size: 20px;
-}
-
-.LightSettingsSubTitle1 {
-  font-size: 20px;
   height: 20px;
   top: -8px;
   position: relative;
